backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
deltalake==1.5.0
fastapi==0.104.1
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.4.3
pandas==3.0.1
psycopg2-binary==2.9.10
pyarrow==23.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

# Development dependencies
ruff==0.5.7
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
